Move the filter types next to the filters
FilterConfig and FilterType live in the XZ module, so they only exist
with the "xz" feature. The sans-I/O decoders are always available and
need them, so move them to the filter module, which is not gated.

The crate root keeps re-exporting both, so nothing changes for users.
